requirements.txt
httpx
pandas
python-dotenv
supabase
openai>=1.40.0
python-dotenv>=1.0.0